updated project structure to use poetry
