+++ pyproject.toml
@@ -1,0 +1,18 @@
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
+
+[tool.poetry]
+name = "ygo_small_world"
+version = "0.1.0"
+description = "A bridge finder for the card 'Small World' in the card game 'Yu-Gi-Oh!'."
+authors = ["Kenneth Allen <kencubed@gmail.com>"]
+license = "MIT"
+
+[tool.poetry.dependencies]
+python = ">=3.8"
+pandas = "^1.0"
+numpy = "^1.20"
+matplotlib = "^3.4"
+requests = "^2.26"
+networkx = "^2.6"
